+++ jarvis_sdk_node/src/sdk/model/hash_algorithm/__test__/bcrypt.test.ts
@@ -1,0 +1,17 @@
+import Bcrypt from '../bcrypt';
+import HashAlgorithm from '../hash_algorithm';
+
+describe('when a new instance is created', () => {
+  let instance: HashAlgorithm;
+
+  beforeEach(() => {
+    instance = new Bcrypt();
+  });
+
+  it('marshals correctly', () => {
+    expect(instance.marshal()).toEqual({
+      oneofKind: 'bcrypt',
+      bcrypt: {},
+    });
+  });
+});
