+++ jarvis_sdk_node/src/sdk/model/hash_algorithm/__test__/hmacmd5.test.ts
@@ -1,0 +1,20 @@
+import HMACMD5 from '../hmacmd5';
+import HashAlgorithm from '../hash_algorithm';
+
+describe('when a new instance is created', () => {
+  const key = Buffer.from('key-value');
+  let instance: HashAlgorithm;
+
+  beforeEach(() => {
+    instance = new HMACMD5(key);
+  });
+
+  it('marshals correctly', () => {
+    expect(instance.marshal()).toEqual({
+      oneofKind: 'hmacMd5',
+      hmacMd5: {
+        key,
+      },
+    });
+  });
+});
