+++ jarvis_sdk_node/src/sdk/model/hash_algorithm/__test__/pbkdf_sha1.test.ts
@@ -1,0 +1,20 @@
+import PBKDFSHA1 from '../pbkdf_sha1';
+import HashAlgorithm from '../hash_algorithm';
+
+describe('when a new instance is created', () => {
+  const rounds = '14';
+  let instance: HashAlgorithm;
+
+  beforeEach(() => {
+    instance = new PBKDFSHA1(rounds);
+  });
+
+  it('marshals correctly', () => {
+    expect(instance.marshal()).toEqual({
+      oneofKind: 'pbkdfSha1',
+      pbkdfSha1: {
+        rounds,
+      },
+    });
+  });
+});
