+++ jarvis_sdk_node/src/sdk/model/hash_algorithm/__test__/sha256.test.ts
@@ -1,0 +1,20 @@
+import SHA256 from '../sha256';
+import HashAlgorithm from '../hash_algorithm';
+
+describe('when a new instance is created', () => {
+  const rounds = '14';
+  let instance: HashAlgorithm;
+
+  beforeEach(() => {
+    instance = new SHA256(rounds);
+  });
+
+  it('marshals correctly', () => {
+    expect(instance.marshal()).toEqual({
+      oneofKind: 'sha256',
+      sha256: {
+        rounds,
+      },
+    });
+  });
+});
