+++ jarvis_sdk_node/src/sdk/model/hash_algorithm/__test__/sha512.test.ts
@@ -1,0 +1,20 @@
+import SHA512 from '../sha512';
+import HashAlgorithm from '../hash_algorithm';
+
+describe('when a new instance is created', () => {
+  const rounds = '14';
+  let instance: HashAlgorithm;
+
+  beforeEach(() => {
+    instance = new SHA512(rounds);
+  });
+
+  it('marshals correctly', () => {
+    expect(instance.marshal()).toEqual({
+      oneofKind: 'sha512',
+      sha512: {
+        rounds,
+      },
+    });
+  });
+});
