+++ jarvis_sdk_node/src/sdk/model/hash_algorithm/__test__/standard_scrypt.test.ts
@@ -1,0 +1,27 @@
+import StandardScrypt from '../standard_scrypt';
+import HashAlgorithm from '../hash_algorithm';
+
+describe('when a new instance is created', () => {
+  let instance: HashAlgorithm;
+
+  beforeEach(() => {
+    instance = new StandardScrypt(
+      'block-size',
+      'derived-key-length',
+      'memory-cost',
+      'parallelization',
+    );
+  });
+
+  it('marshals correctly', () => {
+    expect(instance.marshal()).toEqual({
+      oneofKind: 'standardScrypt',
+      standardScrypt: {
+        blockSize: 'block-size',
+        derivedKeyLength: 'derived-key-length',
+        memoryCost: 'memory-cost',
+        parallelization: 'parallelization',
+      },
+    });
+  });
+});
