+++ jarvis_sdk_node/src/sdk/utils/test_utils.ts
@@ -1,0 +1,15 @@
+export const applicationTokenMock = {
+  appSpaceId: '00000000-0000-4444-8888-000000000000',
+  appAgentId: '00000000-0000-4444-8888-000000000000',
+  endpoint: 'example.com',
+  privateKeyJWK: {
+    kty: 'EC',
+    d: 'd-size',
+    use: 'sig',
+    crv: 'P-256',
+    kid: 'kid-token',
+    x: '2cx4aENdH7GTOj5hQCfvvAqZ3HwpFLfi4mo_nK7D4q0',
+    y: 'eknOKcgBUdaak2ZS2NHCEBWcdoLiZ1u7vCIoWZCs8WI',
+    alg: 'ES256',
+  },
+};
